In this modification:
We use a BinaryWriter to write binary data to the output file.
The header now includes the number of distinct characters in the frequency table and the length of the encoded text.
The encoded text is written directly as binary data to the output file.
This structure allows for a more compact and efficient encoding of the data. The output file will be named <inputFileName>_compressed.bin and will contain the compressed data along with the necessary information for decoding.